Refac: use pytest
