+++ src/lime/__init__.py
@@ -1,3 +1,1 @@
-__package__ = 'lime-python'
-
-from lime.protocol import *
+from .protocol import *
