+++ src/lime/protocol/serializable.py
@@ -10,7 +10,7 @@ class Serializable:
         Transform class properties to json.
 
         Returns:
-            dict
+            dict: the class json representation without private properties
         """
         return {
             self.normalize_key(key): value
@@ -23,9 +23,9 @@ class Serializable:
         Normalize a class property name.
 
         Args:
-            key: (str) property name
+            key (str): property name
 
         Returns:
-            str
+            str: the normalized key
         """
         return key.replace(NODE_KEY_TOKEN, str())
